requirements.txt
django
django-rest-framework
django-oauth-toolkit
